+++ test/esri-loader.spec.js
@@ -4,7 +4,7 @@ describe('esri-loader', function () {
       // remove previously stubbed require function
       delete window.require;
       // esri-loader script has not yet been loaded
-      document.querySelector = jasmine.createSpy('querySelector').and.returnValue(null);
+      spyOn(document, 'querySelector').and.returnValue(null);
     });
     it('isLoaded should be false', function () {
       expect(esriLoader.isLoaded())
@@ -57,6 +57,47 @@ describe('esri-loader', function () {
         expect(context.bootstrapCallback).toHaveBeenCalledTimes(1);
       });
     });
+    describe('when called twice', function () {
+      var scriptEl;
+      var context = {
+        bootstrapCallback: function () {}
+      };
+      describe('w/ a callback', function () {
+        beforeAll(function () {
+          spyOn(document.body, 'appendChild').and.callFake(function (el) {
+            // hold onto script element for assertions below
+            scriptEl = el;
+            // call the onload callback
+            el.onload();
+          });
+          spyOn(document, 'querySelector').and.callFake(function() {
+            console.log('querySelector', scriptEl);
+            return scriptEl;
+          });
+          spyOn(context, 'bootstrapCallback');
+        });
+        it('should pass an error to the callback', function () {
+          esriLoader.bootstrap(context.bootstrapCallback, {
+            url: 'https://js.arcgis.com/3.20'
+          });
+          esriLoader.bootstrap(context.bootstrapCallback, {
+            url: 'https://js.arcgis.com/3.20'
+          });
+          expect(context.bootstrapCallback.calls.argsFor(1)[0].message).toEqual('The ArcGIS API for JavaScript is already loaded.');
+        });
+      });
+      describe('w/o a callback', function () {
+        it('should not throw when called w/o a callback', function () {
+          esriLoader.bootstrap(null, {
+            url: 'https://js.arcgis.com/3.20'
+          });
+          esriLoader.bootstrap(null, {
+            url: 'https://js.arcgis.com/3.20'
+          });
+          expect(1).toEqual(1);
+        });
+      });
+    });
   });
 
   describe('when loading modules', function () {
@@ -72,7 +113,7 @@ describe('esri-loader', function () {
         callback();
       };
       var esriLoaderScript = document.createElement('script');
-      document.querySelector = jasmine.createSpy('querySelector').and.returnValue(esriLoaderScript);
+      spyOn(document, 'querySelector').and.returnValue(esriLoaderScript);
       spyOn(context, 'requireCallback');
       esriLoader.dojoRequire(expectedModuleNames, context.requireCallback);
     });